it's simply implementation of doubly link list
This is a doubly linked list program in C++. Each Node will have a reference pointer to its next as well as previous node.

It maintains an ascending order. So new elements are inserted in the right place.

Reversing A Linked List will reverse the list as well as the property Ascending becomes descending and vice-versa. So Order of various Process :